requirements_test.txt
# Test dependencies for Cudy Router integration
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
